Flatten accounts and relationships into Hunter
