+++ test/hunter/account_test.exs
@@ -14,7 +14,7 @@ defmodule Hunter.AccountTest do
     end)
 
     assert %Account{username: "milmazz", followers_count: 118} =
-             Account.verify_credentials(@conn)
+             Hunter.verify_credentials(@conn)
   end
 
   test "updates authenticated user's credentials with a JSON body" do
@@ -25,7 +25,7 @@ defmodule Hunter.AccountTest do
       respond_with_fixture(conn, "account")
     end)
 
-    assert %Account{username: "milmazz"} = Account.update_credentials(@conn, %{note: "new bio"})
+    assert %Account{username: "milmazz"} = Hunter.update_credentials(@conn, %{note: "new bio"})
   end
 
   test "returns an account" do
@@ -35,7 +35,7 @@ defmodule Hunter.AccountTest do
       respond_with_fixture(conn, "account")
     end)
 
-    assert %Account{username: "milmazz", acct: "milmazz"} = Account.account(@conn, 23_634)
+    assert %Account{username: "milmazz", acct: "milmazz"} = Hunter.account(@conn, 23_634)
   end
 
   test "returns a collection of followers with query params" do
@@ -46,7 +46,7 @@ defmodule Hunter.AccountTest do
       respond_with_fixture(conn, "account", wrap: :list)
     end)
 
-    assert [%Account{username: "milmazz"}] = Account.followers(@conn, 23_634, limit: 1)
+    assert [%Account{username: "milmazz"}] = Hunter.followers(@conn, 23_634, limit: 1)
   end
 
   test "returns a collection of followed accounts with query params" do
@@ -57,7 +57,7 @@ defmodule Hunter.AccountTest do
       respond_with_fixture(conn, "account", wrap: :list)
     end)
 
-    assert [%Account{username: "milmazz"}] = Account.following(@conn, 23_634, limit: 1)
+    assert [%Account{username: "milmazz"}] = Hunter.following(@conn, 23_634, limit: 1)
   end
 
   test "searches for accounts with the q param and a default limit" do
@@ -69,7 +69,7 @@ defmodule Hunter.AccountTest do
       respond_with_fixture(conn, "account", wrap: :list)
     end)
 
-    assert [%Account{username: "milmazz"}] = Account.search_account(@conn, q: "milmazz")
+    assert [%Account{username: "milmazz"}] = Hunter.search_account(@conn, q: "milmazz")
   end
 
   test "returns blocked accounts" do
@@ -79,7 +79,7 @@ defmodule Hunter.AccountTest do
       respond_with_fixture(conn, "account", wrap: :list)
     end)
 
-    assert [%Account{}] = Account.blocks(@conn)
+    assert [%Account{}] = Hunter.blocks(@conn)
   end
 
   test "returns follow requests" do
@@ -89,7 +89,7 @@ defmodule Hunter.AccountTest do
       respond_with_fixture(conn, "account", wrap: :list)
     end)
 
-    assert [%Account{}] = Account.follow_requests(@conn)
+    assert [%Account{}] = Hunter.follow_requests(@conn)
   end
 
   test "returns muted accounts" do
@@ -99,7 +99,7 @@ defmodule Hunter.AccountTest do
       respond_with_fixture(conn, "account", wrap: :list)
     end)
 
-    assert [%Account{}] = Account.mutes(@conn)
+    assert [%Account{}] = Hunter.mutes(@conn)
   end
 
   test "accepts a follow request" do
@@ -109,7 +109,7 @@ defmodule Hunter.AccountTest do
       respond_with_fixture(conn, "relationship")
     end)
 
-    assert %Hunter.Relationship{id: "8039"} = Account.accept_follow_request(@conn, 8039)
+    assert %Hunter.Relationship{id: "8039"} = Hunter.accept_follow_request(@conn, 8039)
   end
 
   test "rejects a follow request" do
@@ -119,7 +119,7 @@ defmodule Hunter.AccountTest do
       respond_with_fixture(conn, "relationship")
     end)
 
-    assert %Hunter.Relationship{id: "8039"} = Account.reject_follow_request(@conn, 8039)
+    assert %Hunter.Relationship{id: "8039"} = Hunter.reject_follow_request(@conn, 8039)
   end
 
   test "returns the accounts that reblogged a status" do
@@ -130,7 +130,7 @@ defmodule Hunter.AccountTest do
       respond_with_fixture(conn, "account", wrap: :list)
     end)
 
-    assert [%Account{username: "milmazz"}] = Account.reblogged_by(@conn, 153_452, limit: 1)
+    assert [%Account{username: "milmazz"}] = Hunter.reblogged_by(@conn, 153_452, limit: 1)
   end
 
   test "returns the accounts that favourited a status" do
@@ -141,7 +141,7 @@ defmodule Hunter.AccountTest do
       respond_with_fixture(conn, "account", wrap: :list)
     end)
 
-    assert [%Account{username: "milmazz"}] = Account.favourited_by(@conn, 153_452, limit: 1)
+    assert [%Account{username: "milmazz"}] = Hunter.favourited_by(@conn, 153_452, limit: 1)
   end
 
   test "API errors raise Hunter.Error" do
@@ -149,6 +149,6 @@ defmodule Hunter.AccountTest do
       respond_with(conn, %{error: "Record not found"}, 404)
     end)
 
-    assert_raise Hunter.Error, fn -> Account.account(@conn, 0) end
+    assert_raise Hunter.Error, fn -> Hunter.account(@conn, 0) end
   end
 end
